Progress in the front end of the crud crm
Queries and Mutations now working, gotta work on the user authentication access. as of now
the user can access all pages without even having an acount which is not suitable for what
we are doing, im planning on adding it to the layout component which encapsules the whole
content of the website but that would need a document read everytime the user changes the
page, imma need to improve this idea further or come up with something new.